# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/invites:
    post:
      operationId: InvitesController_create
      summary: Send a friend invite by email (stub email transport)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteDto"
      responses:
        "201":
          description: ""
      tags:
        - invites
      security:
        - access-token: []
  /api/v1/invites/me:
    get:
      operationId: InvitesController_listMine
      summary: List invites I have sent
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - invites
      security:
        - access-token: []
  /api/v1/invites/preview/{token}:
    get:
      operationId: InvitesController_preview
      summary: Public invite preview for signup
      description: Token is the same value as the `invite` query param on the registration URL.
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - invites
  /api/v1/invites/{id}:
    delete:
      operationId: InvitesController_cancel
      summary: Cancel a pending invite
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "204":
          description: ""
      tags:
        - invites
      security:
        - access-token: []
...
  /api/v1/friendships:
    get:
      operationId: FriendshipsController_listMine
      summary: List my friendships (filterable by status)
      parameters:
        - name: status
          required: false
          in: query
          schema:
            enum:
              - PENDING
              - ACCEPTED
              - BLOCKED
            type: string
      responses:
        "200":
          description: ""
      tags:
        - friendships
      security:
        - access-token: []
    post:
      operationId: FriendshipsController_request
      summary: Send a friend request
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFriendshipDto"
      responses:
        "201":
          description: ""
      tags:
        - friendships
      security:
        - access-token: []
  /api/v1/friendships/friends:
    get:
      operationId: FriendshipsController_listFriends
      summary: List accepted friends with presence and mutual quest count
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - friendships
      security:
        - access-token: []
  /api/v1/friendships/pending:
    get:
      operationId: FriendshipsController_listPending
      summary: List pending incoming friend requests with requester details
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - friendships
      security:
        - access-token: []
  /api/v1/friendships/activity:
    get:
      operationId: FriendshipsController_activity
      summary: Friend activity feed (quest completions, starts, achievements, level-ups, place visits); pass meta.nextCursor as cursor for the next page
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: cursor
          required: false
          in: query
          description: ISO 8601 instant; only events strictly older than this are returned
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - friendships
      security:
        - access-token: []
  /api/v1/friendships/{id}/accept:
    post:
      operationId: FriendshipsController_accept
      summary: Accept a pending friend request
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - friendships
      security:
        - access-token: []
  /api/v1/friendships/block/{userId}:
    post:
      operationId: FriendshipsController_block
      summary: Block another user
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      tags:
        - friendships
      security:
        - access-token: []
  /api/v1/friendships/{id}:
    delete:
      operationId: FriendshipsController_remove
      summary: Remove a friendship
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - friendships
      security:
        - access-token: []
  /api/v1/businesses:
    get:
      operationId: BusinessesController_list
      summary: List businesses (paginated)
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Free-text search query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BusinessDto"
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: "#/components/schemas/PaginationMeta"
      tags:
        - businesses
      security:
        - access-token: []
    post:
      operationId: BusinessesController_create
      summary: Register a new business
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessDto"
      responses:
        "201":
          description: ""
      tags:
        - businesses
      security:
        - access-token: []
  /api/v1/businesses/me:
    get:
      operationId: BusinessesController_getMine
      summary: My business (owner portal)
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - businesses
      security:
        - access-token: []
...
  /api/v1/businesses/me/quests/top:
    get:
      operationId: BusinessesController_getMyTopQuests
      summary: Top quests by visits in the selected period
      parameters:
        - name: period
          required: true
          in: query
          schema:
            $ref: "#/components/schemas/BusinessMetricsPeriod"
        - name: limit
          required: true
          in: query
          schema:
            minimum: 1
            maximum: 50
            default: 10
            type: number
      responses:
        "200":
          description: ""
      tags:
        - businesses
      security:
        - access-token: []
  /api/v1/businesses/me/onboarding:
    post:
      operationId: BusinessesController_startOnboarding
      summary: Mark partner onboarding as started
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - businesses
      security:
        - access-token: []
  /api/v1/businesses/{id}:
    get:
      operationId: BusinessesController_getById
      summary: Get business by id
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - businesses
      security:
        - access-token: []
  /api/v1/notifications:
    get:
      operationId: NotificationsController_list
      summary: List my notifications (paginated)
      parameters:
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: search
          required: false
          in: query
          description: Free-text search query
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationDto"
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: "#/components/schemas/PaginationMeta"
      tags:
        - notifications
      security:
        - access-token: []
  /api/v1/notifications/unread-count:
    get:
      operationId: NotificationsController_unreadCount
      summary: Unread notification count
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationUnreadCountDto"
      tags:
        - notifications
      security:
        - access-token: []
  /api/v1/notifications/{id}/read:
    patch:
      operationId: NotificationsController_markRead
      summary: Mark a notification as read
      parameters:
        - name: id
          required: true
          in: path
          description: CUID identifier
          schema:
            type: string
      responses:
        "204":
          description: ""
      tags:
        - notifications
      security:
        - access-token: []
  /api/v1/notifications/read-all:
    patch:
      operationId: NotificationsController_markAllRead
      summary: Mark all my notifications as read
      parameters: []
      responses:
        "204":
          description: ""
      tags:
        - notifications
      security:
        - access-token: []
...
info:
  title: SideQuest API
  description: Gamified exploration platform - REST API
  version: "1"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Paste the access token returned by /auth/login
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: number
        limit:
          type: number
        total:
          type: number
        totalPages:
          type: number
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
...
    ParticipantSummaryDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: object
        avatarUrl:
          type: object
        level:
          type: number
      required:
        - id
        - username
        - level
...
    CreateInviteDto:
      type: object
      properties:
        email:
          type: string
          example: friend@example.com
      required:
        - email
    CreateFriendshipDto:
      type: object
      properties:
        addresseeId:
          type: string
          description: User id of the addressee
      required:
        - addresseeId
    BusinessDto:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: object
        websiteUrl:
          type: object
        logoUrl:
          type: object
        status:
          type: string
          enum:
            - PENDING_VERIFICATION
            - VERIFIED
            - REJECTED
            - SUSPENDED
        address:
          type: object
        latitude:
          type: object
        longitude:
          type: object
        ownerId:
          type: object
        onboardingStartedAt:
          type: object
          description: Set when the owner starts the partner onboarding flow
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - slug
        - name
        - status
        - createdAt
        - updatedAt
    BusinessMetricsPeriod:
      type: string
      enum:
        - MONTH
        - 7D
        - 30D
        - 90D
    CreateBusinessDto:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        websiteUrl:
          type: string
        logoUrl:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - slug
        - name
    NotificationType:
      type: string
      enum:
        - QUEST_COMPLETED
        - ACHIEVEMENT_UNLOCKED
        - FRIEND_REQUEST_RECEIVED
        - FRIEND_REQUEST_ACCEPTED
        - LEVEL_UP
        - BUSINESS_QUEST_APPROVED
    NotificationDataDto:
      type: object
      properties:
        questId:
          type: string
        xpAwarded:
          type: number
        achievementId:
          type: string
        slug:
          type: string
        friendshipId:
          type: string
        fromUser:
          $ref: "#/components/schemas/ParticipantSummaryDto"
        previousLevel:
          type: number
        newLevel:
          type: number
    NotificationDto:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          $ref: "#/components/schemas/NotificationType"
        title:
          type: string
        body:
          type: object
        data:
          $ref: "#/components/schemas/NotificationDataDto"
        readAt:
          type: object
        createdAt:
          format: date-time
          type: string
      required:
        - id
        - userId
        - type
        - title
        - createdAt
    NotificationUnreadCountDto:
      type: object
      properties:
        count:
          type: number
          description: Notifications with readAt unset
      required:
        - count